+++ app/src/main/java/com/thedeliveryapp/thedeliveryapp/recyclerview/RecyclerViewOrderAdapter.java
@@ -97,7 +97,7 @@ public class RecyclerViewOrderAdapter extends RecyclerView.Adapter<OrderViewHold
             //Use the provided View Holder on the onCreateViewHolder method to populate the current row on the RecyclerView
             status = String.valueOf(order.status.charAt(0));
             price = Integer.toString(order.max_range);
-            charge = Integer.toString(order.deliveryCharge);
+            charge = Float.toString(order.deliveryCharge);
             holder.category.setText(order.category);
             setExpiry(order.expiryDate,order.expiryTime);
             String date = DateFormat.getDateInstance(DateFormat.MEDIUM).format(calendar.getTime());
